[update]JSのimagecontrollerにてPOSTする処理を別コントローラに移動、別コントローラに移動したため、新規投稿ページの画像処理に関して修正, 観光地設定ページのレイアウトを修正
