+++ app/javascript/controllers/index.js
@@ -10,6 +10,9 @@ application.register("hello", HelloController)
 import ImagesController from "./images_controller"
 application.register("images", ImagesController)
 
+import PostController from "./post_controller"
+application.register("post", PostController)
+
 import SelectmenuController from "./selectmenu_controller"
 application.register("selectmenu", SelectmenuController)
