+++ app/views/layouts/application.html.erb
@@ -47,7 +47,7 @@
               新規投稿
             <% end %>
           </li>
-          <li class="nav-item">
+          <li class="nav-item">  
             <%= link_to destroy_user_session_path , method: :delete, class: "btn btn-danger" do %>
               ログアウト
             <% end %>
