+++ app/views/public/places/new.html.erb
@@ -1,29 +1,27 @@
-<div class="container">
+<div class="container-fluid">
   <div class="row">
     <div class="col">
       <%= form_with model: @place, url: post_places_path(@post_id), method: :post do |f| %>
         <div class="form-group">
-          <%= f.label :place_name, "訪れた場所" %><br>
-          <%= f.text_field :place_name %>
-        </div>
-        <div class="form-group">
-          <%= f.label :address, "住所" %><br>
-          <%= f.text_field :address %>
+          <%= f.label :place_name, "訪問地をキーワードで検索" %><br>
+          <%= f.text_field :place_name, placeholder:'キーワードを入力' %>
         </div>
 
         <div class="form-group">
-          <% if @place_num == nil %>
-            <%= f.select :place_num, ["新規追加"] %>
-          <% else %>
-            <%= f.select :place_num, [@place_num,"新規追加"] %>
-          <% end %>
-
+          <%= f.label :comment, "感想やメモ" %><br>
+          <%= f.text_field :comment, placeholder:'初めて勾当台公園へ行った結構広く参加したイベントも楽しかった', size: " 50 x 40 "  %>
         </div>
 
-        <div class="form-group">
-          <%= f.label :comment, "コメント" %><br>
-          <%= f.text_field :comment %>
+
+         <div class="imagedrag col-md-6"
+             data-images-target="drop"
+             data-action="dragover->images#dragOver dragleave->images#dragLeave drop->images#dropImages">
+          <i class="fa-solid fa-circle-plus"></i></br>
+          写真追加
         </div>
+
+        <div data-images-target="preview" class="preview-area mt-3"></div>
+
 
         <%= f.submit "登録" ,class: "btn btn-primary"%>
